Added PDF Project Report helper utilities plus basic Excel Spreadsheet formatters
Note that switching from commercial pdfJet 5.9.0 to free 5.7.5 left one class stranded, dealing with column text formatting, so I need to run another pass tomorrow in order to work out a different way of doing that with that class missing from the most recent free version from 2016. Not sure if this library is still being developed, but it's more fleshed out than the ones that are based on Graphics2D overrides, as it is oriented towards semantic report generation rather than vectorization of application on-screen graphics.